cassandra_read/requirements.txt
kafka-python==2.0.2
cassandra-driver==3.25.0